// src/sdks/systemV2ApiDocs.ts
export interface Org {
  companyId?: number;

  createBy?: number;

  createTime?: Date;

  deleted?: boolean;

  id?: number;

  orgAbbr?: string;

  orgCode?: string;

  orgInfo?: string;

  orgName?: string;

  orgPath?: string;

  orgType?: '0' | '1' | '2' | '3';

  pid?: number;

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;
}
// ...
export interface Role {
  companyId?: number;

  createBy?: number;

  createTime?: Date;

  deleted?: boolean;

  id?: number;

  roleInfo?: string;

  roleName?: string;

  roleType?: '0' | '1';

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;

  linkUser?: boolean;
}
// ...
export interface UserResp {
  adminFlag?: boolean;

  companyId?: number;

  defaultPassword?: boolean;

  deleted?: boolean;

  id?: number;

  lastLoginTime?: Date;

  mobile?: string;

  orgNames?: string;

  realName?: string;

  roleNames?: string;

  updateBy?: number;

  updateByName?: string;

  updateTime?: Date;

  userStatus?: '0' | '1';

  username?: string;
}